feat: add support for path-based rules (#69)
Rules can now define a optional list of `paths` they apply to. `paths`
support the same syntax as entries in `.gitignore`.

It's valid to specify `paths` at any nesting level within the config.

The absense of `paths` makes evaluates a rule for any file as before, so
this is a backwards compatible change.